backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
fastapi==0.136.1
idna==3.13
psycopg2-binary==2.9.12
pydantic==2.13.3
pydantic_core==2.46.3
python-dotenv==1.2.2
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
unicorn==2.1.4
